pyproject.toml
[tool.ruff]
line-length = 88

[tool.poetry]
name = "use-case-solution"
version = "0.1.0"
description = ""
authors = ["phillipmogensen <pbm@math.ku.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
scikit-learn = "^1.3.2"
pandas = "^2.1.3"
numpy = "^1.26.2"
streamlit = "^1.28.2"
ruff = "^0.1.5"
black = "^23.11.0"
polars = "^0.19.13"
pyyaml = "^6.0.1"
matplotlib = "^3.8.2"
clearml = "^1.13.2"
setuptools = "^68.2.2"
scipy = "^1.11.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
